requirements.txt
numpy
scipy
keras
pillow
tensorflow
scikit-learn
matplotlib
scikit-plot

